refactor : addTodo 컴포넌트
